cleaned up app after merge
