+++ view/src/App.js
@@ -1,3 +1,4 @@
+
 import React, { Component } from 'react';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
 import axios from 'axios';
@@ -7,141 +8,13 @@ import Authentication from './components/authentication.js';
 import Inventory from './components/Inventory/Inventory.js';
 
 class App extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      user: '',
-      pass: '',
-      response: '',
-      authRes: '',
-      token: '',
-      isAuth: false,
-      siteID: 1,
-      studyYear: 2017,
-      inventoryList: []
-    };
-    this.handleUser = this.handleUser.bind(this);
-    this.handlePass = this.handlePass.bind(this);
-    this.handleSubmit = this.handleSubmit.bind(this);
-    this.handleSignup = this.handleSignup.bind(this);
-    this.callApi = this.callApi.bind(this);
-    this.saveToken = this.saveToken.bind(this);
-  }
 
-  async componentDidMount() {
-    await this.saveToken();
-    if (localStorage.getItem('tkkn')) {
-      let data = JSON.stringify({
-        token: localStorage.getItem('tkkn')
-      });
-      await this.callApi('/token', data)
-        .then(res => this.setState({ authRes: res.message, isAuth: res.ans }))
-        .catch(err => console.log(err));
-    }
-    axios.get('/flora/inventory/list/retrieve').then(res => {
-      this.setState({ inventoryList: res.data });
-    });
-  }
-
-  callApi = async (route, data) => {
-
-    switch (route) {
-      case '/test':
-        var response = await fetch(route);
-        var body = await response.json();
-
-        if (response.status !== 200) throw Error(body.message);
-
-        return body;
-
-
-      case '/auth/login':
-        var responseLog = await fetch(route, {
-          method: 'POST',
-          headers: {
-            'Accept': 'application/json',
-            'Content-Type': 'application/json',
-          },
-          body: data
-        })
-        var bodyLog = await responseLog.json();
-
-        if (responseLog.status !== 200) throw Error(bodyLog.message);
-
-        return bodyLog;
-
-
-      case '/auth/signup':
-        var responseSign = await fetch(route, {
-          method: 'POST',
-          headers: {
-            'Accept': 'application/json',
-            'Content-Type': 'application/json',
-          },
-          body: data
-        })
-        var bodySign = await responseSign.json();
-
-        if (responseSign.status !== 200) throw Error(bodySign.message);
-
-        return bodySign;
-      case '/token':
-        var tokenRes = await fetch(route, {
-          method: 'POST',
-          headers: {
-            'Accept': 'application/json',
-            'Content-Type': 'application/json',
-          },
-          body: data
-        })
-        var tkbody = await tokenRes.json();
-
-        if (tokenRes.status !== 200) throw Error(tkbody.message);
-
-        return tkbody;
-    }
+  state = {
+    isAuth: false
   };
 
-  handleUser(event) {
-    this.setState({ user: event.target.value });
-  }
-
-  handlePass(event) {
-    this.setState({ pass: event.target.value });
-  }
-
-  handleSubmit(event) {
-    var data = JSON.stringify({
-      user: this.state.user,
-      pass: this.state.pass
-    })
-
-    alert('A name was submitted: ' + this.state.user + ' also a password.. opsie: ' + this.state.pass);
-    this.callApi('/auth/login', data).then(res => {
-      let token = res.token
-      localStorage.setItem('tkkn', token);
-      this.setState({ token: token, authRes: res.message, isAuth: res.ans });
-    });
-    event.preventDefault();
-  }
-
-  saveToken() {
-    if (localStorage.hasOwnProperty('tkkn')) {
-      this.setState({ token: localStorage.getItem('tkkn') });
-    }
-  }
-
-  handleSignup(event) {
-
-    var data = JSON.stringify({
-      user: this.state.user,
-      pass: this.state.pass
-    });
-
-    alert('A name was submitted: ' + this.state.user + ' also a password.. opsie: ' + this.state.pass);
-    this.callApi('/auth/signup', data).then(res => this.setState({ authRes: res.message }));
-    event.preventDefault();
-
+  checkAuth = (status) => {
+    this.setState({ isAuth: status });
   }
 
   render() {
@@ -153,24 +26,12 @@ class App extends Component {
         </header>
         <Router>
           <div>
-            {!this.state.isAuth && <Route exact path="/" render={(props) => <Authentication {...props} authClick={this.authClick} />} />}
-            {this.state.isAuth && <Route exact path="/" render={(props) => <Inventory {...props} inventoryList={this.state.inventoryList} siteID={this.state.siteID} studyYear={this.state.studyYear} />} />}
+            {!this.state.isAuth && <Route exact path="/" render={(props) => <Authentication {...props} checkAuth={this.checkAuth} />} />}
+            {this.state.isAuth && <Route exact path="/" render={(props) => <Inventory {...props} />} />}
           </div>
         </Router>
-        <p className="App-intro">{this.state.response}</p>
-        <p className="App-intro">{this.state.authRes}</p>
-        <form>
-          <label>
-            <input id="user" type='text' value={this.state.value} onChange={this.handleUser} placeholder="username" />
-            <br />
-            <input id="pass" type='password' value={this.state.value} onChange={this.handlePass} placeholder="password" />
-          </label>
-          <br />
-          <input type="submit" value="Submit" onClick={this.handleSubmit} />
-          <input type="submit" value="Signup" onClick={this.handleSignup} />
-        </form>
       </div>
-    );
+    )
   }
 }
